test_agents: Separate base test class
