+++ tests/agents_tests/test_agents.py
@@ -4,9 +4,6 @@ from __future__ import division
 from __future__ import absolute_import
 from future import standard_library
 standard_library.install_aliases()
-import os
-import tempfile
-import unittest
 
 from chainer import optimizers
 from chainer import testing
@@ -14,13 +11,13 @@ import gym
 gym.undo_logger_setup()
 
 from chainerrl import agents
-from chainerrl.envs.abc import ABC
-from chainerrl.experiments.train_agent import train_agent
 from chainerrl import explorers
 from chainerrl import policies
 from chainerrl import q_functions
 from chainerrl import replay_buffer
 from chainerrl import v_function
+
+import basetest_agents
 
 
 def create_stochastic_policy_for_env(env):
@@ -85,51 +82,12 @@ def create_v_function_for_env(env):
     return v_function.FCVFunction(ndim_obs)
 
 
-class _TestAgentInterface(unittest.TestCase):
-
-    def setUp(self):
-        self.env = ABC(discrete=self.discrete,
-                       partially_observable=self.partially_observable,
-                       episodic=self.episodic)
-
-    def create_agent(self, env):
-        raise NotImplementedError()
-
-    def test_save_load(self):
-        a = self.create_agent(self.env)
-        dirname = tempfile.mkdtemp()
-        a.save(dirname)
-        self.assertTrue(os.path.exists(dirname))
-        b = self.create_agent(self.env)
-        b.load(dirname)
-
-    def test_run_episode(self):
-        agent = self.create_agent(self.env)
-        done = False
-        obs = self.env.reset()
-        t = 0
-        while t < 10 and not done:
-            a = agent.act(obs)
-            obs, r, done, info = self.env.step(a)
-            t += 1
-
-    @testing.attr.slow
-    def test_train(self):
-        agent = self.create_agent(self.env)
-        train_agent(
-            agent=agent,
-            env=self.env,
-            steps=2000,
-            outdir=tempfile.mkdtemp(),
-            max_episode_len=10)
-
-
 @testing.parameterize(*testing.product({
     'discrete': [True, False],
     'partially_observable': [False],
     'episodic': [False],
 }))
-class TestA3C(_TestAgentInterface):
+class TestA3C(basetest_agents._TestAgentInterface):
 
     def create_agent(self, env):
         model = agents.a3c.A3CSeparateModel(
@@ -145,7 +103,7 @@ class TestA3C(_TestAgentInterface):
     'partially_observable': [False],
     'episodic': [False],
 }))
-class TestACER(_TestAgentInterface):
+class TestACER(basetest_agents._TestAgentInterface):
 
     def create_agent(self, env):
         model = agents.acer.ACERSeparateModel(
@@ -163,7 +121,7 @@ class TestACER(_TestAgentInterface):
     'partially_observable': [False],
     'episodic': [False],
 }))
-class TestDQN(_TestAgentInterface):
+class TestDQN(basetest_agents._TestAgentInterface):
 
     def create_agent(self, env):
         model = create_state_q_function_for_env(env)
@@ -180,7 +138,7 @@ class TestDQN(_TestAgentInterface):
     'partially_observable': [False],
     'episodic': [False],
 }))
-class TestDoubleDQN(_TestAgentInterface):
+class TestDoubleDQN(basetest_agents._TestAgentInterface):
 
     def create_agent(self, env):
         model = create_state_q_function_for_env(env)
@@ -198,7 +156,7 @@ class TestDoubleDQN(_TestAgentInterface):
     'partially_observable': [False],
     'episodic': [False],
 }))
-class TestNSQ(_TestAgentInterface):
+class TestNSQ(basetest_agents._TestAgentInterface):
 
     def create_agent(self, env):
         model = create_state_q_function_for_env(env)
@@ -220,7 +178,7 @@ class TestNSQ(_TestAgentInterface):
     'partially_observable': [False],
     'episodic': [False],
 }))
-class TestDDPG(_TestAgentInterface):
+class TestDDPG(basetest_agents._TestAgentInterface):
 
     def create_agent(self, env):
         model = agents.ddpg.DDPGModel(
@@ -241,7 +199,7 @@ class TestDDPG(_TestAgentInterface):
     'partially_observable': [False],
     'episodic': [False],
 }))
-class TestPGT(_TestAgentInterface):
+class TestPGT(basetest_agents._TestAgentInterface):
 
     def create_agent(self, env):
         model = agents.ddpg.DDPGModel(
